refactored some package names and introduced interfaces for some objects
