+++ src/main/java/Main.java
@@ -1,8 +1,9 @@
-import dp.conflict.resolver.parse.AnswerObject;
-import dp.conflict.resolver.parse.FactParser;
-import dp.conflict.resolver.parse.NoConflictException;
-import dp.conflict.resolver.tree.CallTree;
-import dp.conflict.resolver.tree.ConflictType;
+import dp.resolver.parse.AnswerObject;
+import dp.resolver.parse.FactBuilder;
+import dp.resolver.parse.exception.NoConflictException;
+import dp.resolver.tree.CallTree;
+import dp.resolver.tree.ConflictType;
+import dp.resolver.tree.Tree;
 
 import java.io.IOException;
 
@@ -14,11 +15,11 @@ public class Main {
         String test = "/Users/fabian/Projects/Sample/runtime_conflict_sample/Project_A/";
         String target = "/Users/fabian/Projects/Sample/fastjson/";
         String curr = "/Users/fabian/Projects/automized-DP-conflict-resolver/automized-dp-conflict-resolver/";
+        String jar = "/Users/fabian/.m2/repository/org/runtime/conflict/Project_B/1.0/Project_B-1.0.jar";
      /*
         String target = "/Users/fabian/Projects/Sample/commons-collections/";
         String target = "/Users/fabian/Projects/Sample/sample_project/";
         String target = "/Users/fabian/Projects/Sample/conflict_sample/";
-        String jar = "/Users/fabian/.m2/repository/org/example/Project_A/1.0/Project_A-1.0.jar";
     */
 
     /*
@@ -31,15 +32,19 @@ public class Main {
         cf.createPNG();
     */
 
+        /*List<ClazzWithMethodsDto> jarClassList = AssistParser.getJarClassList(jar);
+        System.out.println(jarClassList.toString());
+
+        System.exit(0);*/
 
         long time = System.currentTimeMillis();
-        CallTree tree = null;
+        Tree tree = null;
         AnswerObject answer = new AnswerObject();
         try {
             tree = new CallTree(test, answer);
             tree.computeCallTree();
-            FactParser parser;
-            parser = new FactParser(tree.getConflicts(ConflictType.TYPE_2));
+            FactBuilder parser;
+            parser = new FactBuilder(tree.getConflicts(ConflictType.TYPE_1));
             answer.setIDMap(parser.getIdMap());
             answer.solve();
             long currTime = (System.currentTimeMillis() - time) / 1000 / 60;
