+++ src/main/java/dp/api/github/POMJson.java
@@ -1,0 +1,175 @@
+package dp.api.github;
+
+/*********************************
+ Created by Fabian Oraze on 28.10.20
+ *********************************/
+
+public class POMJson {
+
+    private String name;
+    private String path;
+    private String sha;
+    private String size;
+    private String url;
+    private String html_url;
+    private String git_url;
+    private String download_url;
+    private String type;
+    private String content;
+    private String encoding;
+    private Links _links;
+
+    public POMJson() {
+    }
+
+    public String getSize() {
+        return size;
+    }
+
+    public void setSize(String size) {
+        this.size = size;
+    }
+
+    public POMJson(String name, String path, String sha, String size, String url, String html_url, String git_url, String download_url, String type, String content, String encoding, Links _links) {
+        this.name = name;
+        this.path = path;
+        this.sha = sha;
+        this.size = size;
+        this.url = url;
+        this.html_url = html_url;
+        this.git_url = git_url;
+        this.download_url = download_url;
+        this.type = type;
+        this.content = content;
+        this.encoding = encoding;
+        this._links = _links;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        this.path = path;
+    }
+
+    public String getSha() {
+        return sha;
+    }
+
+    public void setSha(String sha) {
+        this.sha = sha;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+
+    public String getHtml_url() {
+        return html_url;
+    }
+
+    public void setHtml_url(String html_url) {
+        this.html_url = html_url;
+    }
+
+    public String getGit_url() {
+        return git_url;
+    }
+
+    public void setGit_url(String git_url) {
+        this.git_url = git_url;
+    }
+
+    public String getDownload_url() {
+        return download_url;
+    }
+
+    public void setDownload_url(String download_url) {
+        this.download_url = download_url;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    public String getEncoding() {
+        return encoding;
+    }
+
+    public void setEncoding(String encoding) {
+        this.encoding = encoding;
+    }
+
+    public Links get_links() {
+        return _links;
+    }
+
+    public void set_links(Links _links) {
+        this._links = _links;
+    }
+
+    public class Links {
+        private String self;
+        private String git;
+        private String html;
+
+        public Links() {
+        }
+
+        public String getSelf() {
+            return self;
+        }
+
+        public void setSelf(String self) {
+            this.self = self;
+        }
+
+        public String getGit() {
+            return git;
+        }
+
+        public void setGit(String git) {
+            this.git = git;
+        }
+
+        public String getHtml() {
+            return html;
+        }
+
+        public void setHtml(String html) {
+            this.html = html;
+        }
+
+        public Links(String self, String git, String html) {
+            this.self = self;
+            this.git = git;
+            this.html = html;
+        }
+
+    }
+}
