--- src/main/java/dp/conflict/api/github/JSONLoader.java
@@ -1,20 +1,0 @@
-package dp.conflict.api.github;
-
-/*********************************
- Created by Fabian Oraze on 29.10.20
- *********************************/
-
-public class JSONLoader {
-
-
-    /*
-        String query = "https://api.github.com/repos/spotify/docker-maven-plugin/contents/pom.xml";
-        URL url = new URL(query);
-        ObjectMapper objectMapper = new ObjectMapper();
-        POMJson pomJson = objectMapper.readValue(url, POMJson.class);
-
-        System.out.println(pomJson.getContent());
-
-        Base64.Decoder decoder = Base64.getDecoder();
-     */
-}
