--- src/main/java/dp/conflict/resolver/parse/exception/NoConflictException.java
@@ -1,9 +1,0 @@
-package dp.conflict.resolver.parse.exception;
-
-/*********************************
- Created by Fabian Oraze on 22.01.21
- *********************************/
-
-public class NoConflictException extends Exception {
-
-}
