+++ src/main/java/dp/resolver/parse/exception/NoConflictException.java
@@ -1,0 +1,9 @@
+package dp.resolver.parse.exception;
+
+/*********************************
+ Created by Fabian Oraze on 22.01.21
+ *********************************/
+
+public class NoConflictException extends Exception {
+
+}
